NP-49503 Show only institutions with service center uri (#7675)
* Show only institutions with service center uri

* Rename variable

* Simplify check

* Show loading state

* Avoid too long function

* Reduce component size
