+++ src/pages/public_registration/details/DetailsPanel.tsx
@@ -8,8 +8,6 @@ import {
   DialogTitle,
   Divider,
   IconButton,
-  Link,
-  Skeleton,
   styled,
   Typography,
 } from '@mui/material';
@@ -17,25 +15,17 @@ import { visuallyHidden } from '@mui/utils';
 import { useQueries } from '@tanstack/react-query';
 import { useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { fetchOrganization } from '../../api/cristinApi';
-import { useFetchCustomers } from '../../api/hooks/useFetchCustomers';
-import { useFetchPerson } from '../../api/hooks/useFetchPerson';
-import { ContributorName } from '../../components/ContributorName';
-import { OpenInNewLink } from '../../components/OpenInNewLink';
-import { ConfirmedAffiliation, Contributor, ContributorRole } from '../../types/contributor.types';
-import { Organization } from '../../types/organization.types';
-import { dataTestId } from '../../utils/dataTestIds';
-import { getTopLevelOrganization, getUniqueOrganizations } from '../../utils/institutions-helpers';
-import { getLanguageString } from '../../utils/translation-helpers';
+import { fetchOrganization } from '../../../api/cristinApi';
+import { OpenInNewLink } from '../../../components/OpenInNewLink';
+import { ConfirmedAffiliation, Contributor, ContributorRole } from '../../../types/contributor.types';
+import { Organization } from '../../../types/organization.types';
+import { dataTestId } from '../../../utils/dataTestIds';
+import { getTopLevelOrganization, getUniqueOrganizations } from '../../../utils/institutions-helpers';
+import { ContactPersonRow } from './ContactPersonRow';
+import { InstitutionsServiceCenterOverview } from './InstitutionsServiceCenterOverview';
 
 const StyledList = styled('ul')({
   padding: 0,
-});
-
-const StyledListItem = styled('li')({
-  marginBottom: '1rem',
-  marginLeft: 0,
-  listStyleType: 'none',
 });
 
 interface DetailsPanelProps {
@@ -66,10 +56,6 @@ export const DetailsPanel = ({ contributors }: DetailsPanelProps) => {
   const topLevelOrgs = affiliationQueries
     .map((affiliationQuery) => (affiliationQuery.data ? getTopLevelOrganization(affiliationQuery.data) : null))
     .filter(Boolean) as Organization[];
-  const institutions = getUniqueOrganizations(topLevelOrgs);
-
-  const customersData = useFetchCustomers({ enabled: institutions.length > 0, staleTime: 1_800_000 }); // Cache for 30 minutes
-  const customers = customersData.data?.customers ?? [];
 
   return (
     <Box sx={{ display: 'flex', flexDirection: 'column', p: '1rem', bgcolor: 'secondary.main', gap: '0.5rem' }}>
@@ -102,31 +88,7 @@ export const DetailsPanel = ({ contributors }: DetailsPanelProps) => {
         </IconButton>
 
         <DialogContent sx={{ display: 'flex', flexDirection: 'column', gap: '1rem' }}>
-          {institutions.length > 0 && (
-            <div>
-              <Typography variant="h2" gutterBottom>
-                {t('institutions_service_support')}
-              </Typography>
-              <StyledList>
-                {institutions.map((institution) => {
-                  const serviceCenterUri = customers.find(
-                    (customer) => customer.cristinId === institution.id
-                  )?.serviceCenterUri;
-
-                  return (
-                    <StyledListItem key={institution.id}>
-                      <Typography>{getLanguageString(institution.labels)}</Typography>
-                      {serviceCenterUri ? (
-                        <OpenInNewLink href={serviceCenterUri}>{serviceCenterUri}</OpenInNewLink>
-                      ) : (
-                        <Typography fontStyle="italic">{t('no_service_center')} </Typography>
-                      )}
-                    </StyledListItem>
-                  );
-                })}
-              </StyledList>
-            </div>
-          )}
+          <InstitutionsServiceCenterOverview institutions={getUniqueOrganizations(topLevelOrgs)} />
 
           {contactPersons.length > 0 && (
             <div>
@@ -170,46 +132,5 @@ export const DetailsPanel = ({ contributors }: DetailsPanelProps) => {
         </DialogContent>
       </Dialog>
     </Box>
-  );
-};
-
-interface ContactPersonRowProps {
-  contributor: Contributor;
-}
-
-const ContactPersonRow = ({ contributor }: ContactPersonRowProps) => {
-  const id = contributor.identity.id ?? '';
-  const personQuery = useFetchPerson(id);
-  const person = personQuery.data;
-
-  return (
-    <StyledListItem>
-      {personQuery.isFetching ? (
-        <>
-          <Skeleton width="10rem" />
-          <Skeleton width="10rem" />
-        </>
-      ) : (
-        <>
-          <ContributorName
-            id={id}
-            name={contributor.identity.name}
-            hasVerifiedAffiliation={
-              !!contributor.affiliations?.some((affiliation) => affiliation.type === 'Organization')
-            }
-          />
-          {person?.contactDetails?.email && (
-            <Box sx={{ display: 'flex', gap: '0.5rem', alignItems: 'center' }}>
-              <MailOutlineIcon />
-              <Link
-                data-testid={dataTestId.registrationLandingPage.detailsTab.emailLink(id)}
-                href={`mailto:${person.contactDetails.email}`}>
-                {person.contactDetails.email}
-              </Link>
-            </Box>
-          )}
-        </>
-      )}
-    </StyledListItem>
   );
 };
