+++ src/translations/nbTranslations.json
@@ -2276,6 +2276,5 @@
   "view_contact_info": "Se kontaktinformasjon",
   "institution_support": "Institusjonens brukerstøtte",
   "institutions_service_support": "Institusjonenes brukerstøtte",
-  "no_service_center": "Ingen brukerstøttepunkt oppgitt.",
   "vocabulary_missing": "Institusjonen har ikke lagt til vokabular."
 }
